Struct canvas_finalproject::server::CanvasServer

pub struct CanvasServer { /* fields omitted */ }

A struct that contains implementation for behavior associated with the entire server, like creating ClientHandler events when a client tries to establish a connection initially.

Methods

impl CanvasServer

pub fn new(canvas: Canvas) -> Self

Creates a new server object, given a single canvas instance. The canvas is moved in since there is a shared canvas cross clients.